+++ components/ProfilePage/EditProfileBtn.tsx
@@ -1,0 +1,5 @@
+export default function EditProfileBtn(props: any) {
+    return (
+        <button className="uppercase transition  bg-gray-800 text-sm text-white hover:bg-gray-600 px-4 py-2 rounded-md ">Edit Profile </button>
+    );
+}
